+++ project-base/src/Shopsys/ShopBundle/DataFixtures/Base/MailTemplateDataFixture.php
@@ -74,6 +74,19 @@ class MailTemplateDataFixture extends AbstractReferenceFixture
             . 'Log in page: {login_page}';
 
         $this->createMailTemplate($manager, MailTemplate::REGISTRATION_CONFIRM_NAME, $mailTemplateData);
+
+        $mailTemplateData->subject = 'Display personal information - {domain}';
+        $mailTemplateData->body = 'Dear customer, <br /><br />'
+            . '<br />
+            based on your email {e-mail}, we are sending you a link to your personal details. By clicking on <br/> 
+            the link below, you will be taken to a page listing all the personal details which we have on your <br/> 
+            e-mail in our e-shop records {domain}.<br/>
+            Click here for personal information - {url} <br/>
+            <br/>
+            Regards <br/>
+            team {domain} <br/>';
+
+        $this->createMailTemplate($manager, MailTemplate::PERSONAL_DATA_ACCESS_NAME, $mailTemplateData);
     }
 
     private function createMailTemplate(
@@ -81,7 +94,18 @@ class MailTemplateDataFixture extends AbstractReferenceFixture
         $name,
         MailTemplateData $mailTemplateData
     ) {
-        $mailTemplate = new MailTemplate($name, Domain::FIRST_DOMAIN_ID, $mailTemplateData);
+        $repository = $manager->getRepository(MailTemplate::class);
+        $mailTemplate = $repository->findOneBy([
+            'name' => $name,
+            'domainId' => Domain::FIRST_DOMAIN_ID,
+        ]);
+
+        if ($mailTemplate === null) {
+            $mailTemplate = new MailTemplate($name, Domain::FIRST_DOMAIN_ID, $mailTemplateData);
+        } else {
+            $mailTemplate->edit($mailTemplateData);
+        }
+
         $manager->persist($mailTemplate);
         $manager->flush($mailTemplate);
     }