+++ project-base/src/Shopsys/ShopBundle/Model/Gdpr/PersonalDataAccessRequestData.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Gdpr;
+
+class PersonalDataAccessRequestData
+{
+
+    /**
+     * @var \DateTimeImmutable
+     */
+    public $createAt;
+
+    /**
+     * @var string
+     */
+    public $email;
+
+    /**
+     * @var string
+     */
+    public $hash;
+
+    /**
+     * @var int
+     */
+    public $domainId;
+}